safari-600.7-branch/LayoutTests/inspector-obsolete/console/console-nested-group-expected.txt

CONSOLE MESSAGE: line 11: Message inside inner group
CONSOLE MESSAGE: line 14: Message that must not be in any group
CONSOLE MESSAGE: line 21: Message inside third group
Tests that console.group/groupEnd messages won't be coalesced. Bug 56114. Bug 63521.

outer group console-nested-group.html:9 console-message console-log-level console-group-title
inner group console-nested-group.html:10 console-message console-log-level console-group-title
Message inside inner group console-nested-group.html:11 console-message console-log-level
 console-message console-log-level
 console-message console-log-level
Message that must not be in any group console-nested-group.html:14 console-message console-log-level
One of several groups which shouldn't be coalesced. console-nested-group.html:19 console-message console-log-level console-group-title
One of several groups which shouldn't be coalesced. console-nested-group.html:19 console-message console-log-level console-group-title
One of several groups which shouldn't be coalesced. console-nested-group.html:19 console-message console-log-level console-group-title
Message inside third group console-nested-group.html:21 console-message console-log-level
 console-message console-log-level
 console-message console-log-level
 console-message console-log-level
